+++ AutoMl/src/V1beta1/resources/prediction_service_descriptor_config.php
@@ -10,7 +10,7 @@ return [
                     'initialPollDelayMillis' => '500',
                     'pollDelayMultiplier' => '1.5',
                     'maxPollDelayMillis' => '5000',
-                    'totalPollTimeoutMillis' => '300000',
+                    'totalPollTimeoutMillis' => '86400000',
                 ],
             ],
         ],