+++ TextToSpeech/src/V1/SynthesizeSpeechResponse.php
@@ -17,7 +17,8 @@ class SynthesizeSpeechResponse extends \Google\Protobuf\Internal\Message
 {
     /**
      * The audio data bytes encoded as specified in the request, including the
-     * header (For LINEAR16 audio, we include the WAV header). Note: as
+     * header for encodings that are wrapped in containers (e.g. MP3, OGG_OPUS).
+     * For LINEAR16 audio, we include the WAV header. Note: as
      * with all bytes fields, protobuffers use a pure binary representation,
      * whereas JSON representations use base64.
      *
@@ -33,7 +34,8 @@ class SynthesizeSpeechResponse extends \Google\Protobuf\Internal\Message
      *
      *     @type string $audio_content
      *           The audio data bytes encoded as specified in the request, including the
-     *           header (For LINEAR16 audio, we include the WAV header). Note: as
+     *           header for encodings that are wrapped in containers (e.g. MP3, OGG_OPUS).
+     *           For LINEAR16 audio, we include the WAV header. Note: as
      *           with all bytes fields, protobuffers use a pure binary representation,
      *           whereas JSON representations use base64.
      * }
@@ -45,7 +47,8 @@ class SynthesizeSpeechResponse extends \Google\Protobuf\Internal\Message
 
     /**
      * The audio data bytes encoded as specified in the request, including the
-     * header (For LINEAR16 audio, we include the WAV header). Note: as
+     * header for encodings that are wrapped in containers (e.g. MP3, OGG_OPUS).
+     * For LINEAR16 audio, we include the WAV header. Note: as
      * with all bytes fields, protobuffers use a pure binary representation,
      * whereas JSON representations use base64.
      *
@@ -59,7 +62,8 @@ class SynthesizeSpeechResponse extends \Google\Protobuf\Internal\Message
 
     /**
      * The audio data bytes encoded as specified in the request, including the
-     * header (For LINEAR16 audio, we include the WAV header). Note: as
+     * header for encodings that are wrapped in containers (e.g. MP3, OGG_OPUS).
+     * For LINEAR16 audio, we include the WAV header. Note: as
      * with all bytes fields, protobuffers use a pure binary representation,
      * whereas JSON representations use base64.
      *